refactor(vita49): unify receiver output sinks
Route HDF5 and VITA through the same receiver sink contract, moving
HDF5 persistence into a serial sink instead of the DSP finalizer.

Remove backend-specific live-output branching from the simulation engine
and require finalizers to receive their sink from the caller.

Replace fine pacing yields with CPU pause spins for tighter UDP timing.
