+++ packages/libfers/src/processing/finalizer.h
@@ -49,35 +49,6 @@ namespace processing
 							 std::span<const ComplexType> samples, std::uint64_t sample_start,
 							 std::shared_ptr<const core::OutputFileMetadata> file_metadata = nullptr);
 
-	/// HDF5 implementation of the receiver output sink contract.
-	class Hdf5OutputSink final : public core::ReceiverOutputSink
-	{
-	public:
-		Hdf5OutputSink(std::string output_dir,
-					   std::shared_ptr<core::OutputMetadataCollector> metadata_collector = nullptr);
-		~Hdf5OutputSink() override;
-
-		Hdf5OutputSink(const Hdf5OutputSink&) = delete;
-		Hdf5OutputSink& operator=(const Hdf5OutputSink&) = delete;
-
-		void initializeRun(const core::OutputConfig& config, std::string simulation_name) override;
-		std::uint32_t registerStream(const core::ReceiverStreamDescriptor& stream) override;
-		void openStream(std::uint32_t stream_id, RealType first_sample_time) override;
-		void submitBlock(const core::ReceiverSampleBlock& block) override;
-		void emitContextHeartbeat(RealType simulation_time) override;
-		void closeStream(std::uint32_t stream_id) override;
-		core::OutputStats finalize() override;
-
-	private:
-		struct Impl;
-		std::unique_ptr<Impl> _impl;
-	};
-
-	/// Creates the default HDF5 receiver output sink.
-	[[nodiscard]] std::unique_ptr<core::ReceiverOutputSink>
-	makeHdf5OutputSink(std::string output_dir,
-					   std::shared_ptr<core::OutputMetadataCollector> metadata_collector = nullptr);
-
 	/**
 	 * @brief The main function for a dedicated pulsed-mode receiver finalizer thread.
 	 *
